Split 1k+ line ec2 module to multiple files.
The ec2 module is super large, making it hard to navigate.
This commit splits it into multiple files, one per sync method (related resource types). This allows for updated method names that are consistent between the different files. Again making in easier to navigate and recognise familiar patterns.
The naming changes are:
- get_XYZ -> get_from_aws
- load_XYZ -> update_graph
- cleanup_XYZ -> cleanup_graph