Fix query for gcp firewall ingress deny rules.
The bug happens whenever there are existing rules in both `fw['transformed_allow_list']` and in `fw['transformed_deny_list']`. In that case, both of the sub-queries creating the realtionships `[r:ALLOWED_BY]` and `[r:DENIED_BY]` would have been added to the main `query` string, causing the error mentioned in #183 .

I haven't tested this yet, but I believe this should fix it.