Migrate to Neo4j 4.x
This commit makes a few small types of changes to add support for Neo4j
4.x, but they are numerous throughout the codebase.

Firstly, the Neo4j Python Driver is updated to the 4.x series,
specifically the latest 4.2 driver. With it, the neobolt dependency is
dropped as it is no longer required due to Exception classes being
included in the driver package nowadays.

Secondly, all Cypher templates are updated to use the new parameter
syntax of $-prefixed naming instead of the previous {} syntax. This
small syntax change results a LOT of the codebase being touched for
very small changes.

Note: As of the time of this commit, it was required to manually install
version 1.25 of the google-api-core package as the 1.26-dev version
being installed was breaking test execution. No changes were made to
setup.py to accomdate this.