Move GH outside-collab cleanup job to correct place (#1187)
Previously GitHub outside collaborators were cleaned up by the GH user
sync. This is incorrect behavior because outside collaborators are
actually synced in by the GH repo sync. This can be problematic if the
GH sync exhausts API quota and crashes in the middle, leaving the graph
in an inconsistent state.

This quick patch makes it so that the outside collaborators are properly
synced in and cleaned up by the GH repo sync.